create modules and export in commonjs to clean up controller file
