+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "express-intro",
+  "type": "commonjs",
   "version": "1.0.0",
   "description": "\"an intro app for express\"",
   "main": "index.js",
